refactor(graduate): domain에 따른 로직 entity 레이어로 분리
