+++ apps/graduate/src/admin/entities/graduation-approval/api/index.ts
@@ -1,0 +1,1 @@
+export * from './updateGraduationBatchApproval';
