+++ apps/graduate/src/admin/entities/graduation-approval/model/index.ts
@@ -1,0 +1,2 @@
+export * from './useGraduationApproval';
+export * from './useGraduationBatchApproval';
