+++ apps/graduate/src/admin/entities/graduation-users/api/removeGraduationUser.ts
