+++ apps/graduate/src/admin/entities/graduation-users/api/removeGraduationUsersBatch.ts
