+++ apps/graduate/src/admin/entities/graduation-users/model/index.ts
@@ -1,0 +1,2 @@
+export * from './useFetchGraduationUsers';
+export * from './useRemoveGraduationUsers';
