+++ apps/graduate/src/admin/entities/graduation-users/model/useFetchGraduationUsers.ts
@@ -5,8 +5,8 @@ import { KEYS } from '~/shared/constants';
 import {
   type FetchGraduationUsersParams,
   type GraduationUserSummaryPageResponse,
-} from '~/admin/shared/api';
-import { fetchGraduationUsers } from '~/admin/shared/api';
+} from '~/admin/entities/graduation-users/api';
+import { fetchGraduationUsers } from '~/admin/entities/graduation-users/api';
 
 export function useFetchGraduationUsers(params: FetchGraduationUsersParams) {
   return useQuery<GraduationUserSummaryPageResponse>({
