+++ apps/graduate/src/admin/shared/api/index.ts
@@ -1,9 +1,4 @@
 export * from './fetchAdminUsers';
-export * from './fetchGraduationUsers';
-export * from './fetchGraduationUsersExcel';
-export * from './removeGraduationUser';
-export * from './removeGraduationUsersBatch';
-export * from './updateGraduationUsersBatchApprove';
 export * from './schedule';
 export * from './notice';
 export * from './file';
