+++ apps/graduate/src/admin/shared/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from './admin-table';
-export * from './student';
 
 export { useScheduleList } from './useScheduleList';
 export { useScheduleDetail } from './useScheduleDetail';
