--- apps/graduate/src/admin/shared/hooks/student/index.ts
@@ -1,4 +1,0 @@
-export { useFetchGraduationUsers } from './useFetchGraduationUsers';
-export { useApproveGraduationUsers } from './useApproveGraduationUsers';
-export { useRemoveGraduationUsers } from './useRemoveGraduationUsers';
-export { useUpdateGraduationUsersBatchApprove } from './useUpdateGraduationUsersBatchApprove';
