--- apps/graduate/src/client/shared/hooks/index.ts
@@ -1,1 +1,0 @@
-// Client-specific hooks will be added here
