+++ apps/graduate/src/shared/utils/downloadGraduationUsersExcel.ts
@@ -1,6 +1,6 @@
 import { GraduationType } from '../constants';
 
-import { fetchGraduationUsersExcel } from '~/admin/shared/api';
+import { fetchGraduationUsersExcel } from '~/admin/entities/graduation-users/api';
 
 const DEFAULT_FILENAME = '졸업_대상자_목록.xlsx';
